// ticket-monorepo/apps/web-wallet/app/wallet/new-upg/enums/page.enum.ts
export enum PageEnum {
  TERMS_AND_CONDITIONS = 'TERMS_AND_CONDITIONS',
  PAYMENT_METHOD = 'PAYMENT_METHOD',
  CASH_IN_AND_PAY = 'CASH_IN_AND_PAY',
  ICP_INVALID_AMOUNT = 'ICP_INVALID_AMOUNT',
  ICP_VALID_AMOUNT = 'ICP_VALID_AMOUNT',
  CPOTP = 'CPOTP',
  CPPIN = 'CPPIN',
  WALLET_PAY = 'WALLET_PAY',
  OTP = 'OTP',
  PIN = 'PIN',
  CASH_IN_REDIRECT = 'CASH_IN_REDIRECT',
  CASH_OUT = 'CASH_OUT'
}